Separate todolist get test to __test__/todolist #13
