+++ server/src/services/__test__/todolist.spec.ts
@@ -93,71 +93,4 @@ describe('TodolistModule', () => {
       }
     })
   })
-
-  describe('getTodolists', () => {
-    it('should return Todolist type array', async () => {
-      const mockTodolist: Todolist[] = [
-        {
-          id: '1',
-          category: new Category(),
-          categoryId: '1',
-          order: 0,
-          title: 'mock todolist title 1',
-          checked: false,
-          createdAt: new Date(),
-          updatedAt: new Date()
-        },
-        {
-          id: '2',
-          category: new Category(),
-          categoryId: '1',
-          order: 1,
-          title: 'mock todolist title 2',
-          checked: false,
-          createdAt: new Date(),
-          updatedAt: new Date()
-        }
-      ]
-
-      jest.spyOn(service, 'getTodolists').mockResolvedValue({
-        data: mockTodolist,
-        total: mockTodolist.length
-      })
-
-      const result = (await controller.getTodolists()).data
-      const [first, second] = result
-
-      expect(first.id).toBe('1')
-      expect(second.id).toBe('2')
-    })
-
-    it('should return empty array', async () => {
-      const mockTodolist: Todolist[] = []
-
-      jest.spyOn(service, 'getTodolists').mockResolvedValue({
-        data: mockTodolist,
-        total: mockTodolist.length
-      })
-
-      const result = (await controller.getTodolists()).total
-      expect(result).toBe(0)
-    })
-
-    it('when request query with dto that not matched', async () => {
-      const request = {
-        query: {
-          checked: 'tru'
-        }
-      }
-
-      const typiaError = await checkRequestValidate(GetTodolistCheckedValidator, request)
-
-      try {
-        new TypiaExceptionHandler(typiaError.response).handleValidationError()
-      } catch (err) {
-        expect(err).toBeInstanceOf(BadRequestException)
-        expect(err.response.message).toBe(`Received unexpected data '(\"false\" | \"true\" | undefined)' [INVALID QUERY DATA ERROR]`)
-      }
-    })
-  })
 })
